+++ Core/Render/OGL/EMULATOR_RENDER_OGL.C
@@ -1031,10 +1031,10 @@ void Emulator_WaitForTimestep(int count)
 {
 	if (g_swapInterval > 0)
 	{
-		int delta = g_swapTime + FIXED_TIME_STEP * count - Emulator_GetTicks();
+		int64_t delta = g_swapTime + FIXED_TIME_STEP * count - Emulator_GetTicks();
 
 		if (delta > 0) {
-			SDL_Delay(delta);
+			SDL_Delay((unsigned int)delta);
 		}
 	}
